feat : ChallengeCell init
