+++ KnockKnock-iOS/Extension-Utilities/Foundation/Int+Ext.swift
@@ -1,0 +1,8 @@
+//
+//  Int+Ext.swift
+//  KnockKnock-iOS
+//
+//  Created by sangwon yoon on 2022/03/06.
+//
+
+import Foundation
